move Notfound into common and update path.
