+++ src/components/common/NotFound.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { Link } from 'react-router-dom';
-import JumbotronWrapper from './common/JumbotronWrapper';
+import JumbotronWrapper from './JumbotronWrapper';
 
 const NotFound = (props) => (
 	<JumbotronWrapper {...props.jumbotronProps}>
