+++ src/components/common/NotFound.stories.js
